SQL setup (local & prod)
This sets up an updated bitnami SQL chart in both the
local and production environments.

This follows the secret pattern of
TF -> TF State & K8s Secrets
And then
K8s Secrets -> Services

Right now this leaves the services in production already
deployed that expect SQL access hanging, and that will
be fixed in another patch.

This also set ups adminer, using the most popular chart,
that was previously used by wbstack for local development.

This also modifies k8s-shell bin to make it easier for
all services. And adds an SQL example.