Makefile: make local and staging commands easier
This introduced a few new commands aimed at making
syncing things in the 2 existing deployments easier.
This includes terraform and also helmfile.

This potentially highlights something that we may want
to change, which is that terraform is for the whole of our
google environment right now, vs helm which is per environment.