TF: Create wikibase.cloud cluster
N.B. This has been applied to the cluster but until we apply
the helmfile definitions one small part will not work.

Specifically the cert-manager namespace is missing which means that
creating the secret for managing it fails.

We must re-run `terraform apply` after the namespace has been made
by helm.