Fix git not match cluster (#183)
* TF: Fix production git to look like staging

This uses a technique to refer to historic versions of modules
by referencing a specific git hash. This could cause a problem
for users who are unable to access this git repository over ssh
(e.g. they are missing credentials or github is down)

* k8s: Remove backup user sql creation from production

This hasn't yet been deployed so lets remove it for now