tf: Adjust data flow for smtp password (#186)
Due to the constraints of the mailgun resource we ignore
changes on the smtp_password. However, under the hood this
this is stored as a blank value in the terraform state.

This used propagates to the k8s secret if the k8s secret has to be
recreated for some reason which results in email failures.

This patch adjusts the flow of data so we directly use the
random_secret resource to create the k8s secret.