tf: monitoring remove use of locals
This removes the use of locals in this module.
Generally they are best used very sparingly[1].

Also in this case using them removed a dependency that
can prevent the provider from cleanly applying the state.
This was the case when I applied the preceeding state;
it required a double application.

Since the alert policy can only be using a metric that
already exists the provider needs to know the order that the
resources need to be created. This is normally hinted to TF
by using attributes of those resources[2].

[1] https://www.terraform.io/language/values/locals#when-to-use-local-values
[2] https://learn.hashicorp.com/tutorials/terraform/dependencies