k8s: Increase memory for elasticsearch
It seems like the very slow / bad performance of ES node such that it is
unable to become ready may be due to memory exhaustion.

4GB of memory for the pod, half of which is allocated to the heap
seems like a reasonable place to start.

This also "Guaranteed QoS" for memory by setting pod request and
limit to be equal.

These numbers and the QoS strategy were derived from the examples in the
elastic cloud on k8s guide [1].

[1] https://www.elastic.co/guide/en/cloud-on-k8s/current/k8s-managing-compute-resources.html