Local: Rotate root sql password
This commit adds a new sql root password to tf and saves it in the existing k8s secret.
This commit creates a new k8s secret to save the existing tf password.
It also adds a job to be submited using 'kubectl apply -f' that resets the old password
to the new password on the primary mysql server.

Bug:T321051