tf: uptime for CirrusSearch backed MediaWiki api
Add uptime check and alert to uptime module for a
mediawiki api endpoint that depends on being able to
contact elasticsearch.

This endpoint was chosen because it appears to always
contact elasticsearch in order to obtain the mapping.

In order to determine success or failure we have to inspect
the actual JSON response from the api since even errors return
an HTTP status 200.

Doing this has resulted in exposing more of the underlying
resources to the local configuration and added to the
complexity of reading the resource and the module variables.
However, it does ensure consistency with our other uptime checks
and alerts.

This commit doesn't see this module actually being used