k8s:local upgrade kube-prometheus-stack Chart to 56.6.2
This updates the prometheus stack version through
many major versions.

It also updates the prometheus operator CRDs that
kube-prometheus-stack relies on from 2.0.0 to 9.0.0

While the README[1] lists multiple manual actions required
for the kube-prometheus-stack update it appears that
the only ones that apply to us are the CRD updates.

I also tested if the CRD updates appear to be backwards compatible
and to a very cursory test the older version of kube-prometheus-stack
still works and can be fully removed and reapplied to the cluster.

[1] https://github.com/prometheus-community/helm-charts/tree/main/charts/kube-prometheus-stack