feature(staging): add additional platform summary metrics
This follows the pattern of ADR 12 and introduces a "next" version
of the monitoring module.

This next version of the monitoring module contains the mostly renamed
not_edited_last_90_days and edited_last_90_days as well as the new
total_properties_count and total_items_count as default log based
metrics to be created.

This new module is sourced and the overriding of
platform_summary_metrics variable is removed since
it is no longer necessary

Bug: T364452
Bug: T314481