feature(staging): change metrics on platform summary dashboard
Adds total items and properties scorecards.
Adjusts active and inactive to edited/not-edited in 90 days
graphs.
Adjust legend widget for clarity
Add human readable legend aliases to charts

Preview what this dashboard would look like at
https://console.cloud.google.com/monitoring/dashboards/builder/0a7f689a-f7a0-46b8-8dc6-01d28e4e6741

after merging this PR the copied dashboard should be deleted.

By factoring out the dashboard json into a standalone file it
is then trivial to copy paste the json of the copied from the UI
(and adjust the name to not imply it is a copy)

Bug: T364452
Bug: T314481