chore(mediawiki): enable HPA for MW in local and staging
Uses the chart introduced in wbstack/charts#145

This commit also adds a value file for future use in production.
The max value for local and staging are set to 2 to keep
unnecessary scaling to a minimum while also enabling us to
see the effect.

It is set to 5 for production because this gives some headroom
above the currently defined 3.

Bug: T368360