redis-2: downgrade chart to version 19.6.4
This patch downgrades the redis-2 application on local and staging
it preemptively sets the version that would be set on production
too.

This was done because 19.6.4 appears to be the most recent version of
of the chat that ships a redis version (7.2.5) that's still available under
a BSD license.

See: https://redis.io/blog/what-redis-license-change-means-for-our-managed-service-providers/

Bug: T380448