prod: bump argocd dex memory limit to match staging
The argocd-dex-server is continually crashing and restarting with error 139
in production. There are a lot of logs coming form this pod that may be related to
this.

Unfortunately it had less memory request and limit than staging

This patch hopes to resolve the crashing issue but also to correct the discrepancy
that staging has more resources than production; this should probably never happen.

Bug: T380029