prod(tofu): refine production-site-request-count
This copies the staging log based metric

This commit even further refines the log based metric.
Specifically it now isolates the it to only contain
wbaas-3 metrics.

Secondly it tightens the label extraction. This is specifically
to target the edge case of things like urls included in user agent
strings

Bug: T385969