prod(tf): add node pool without highthroughput logging (#2074)
This patch adds a new identical node pool to production without highthroughput logging.

We have very rarely in the last 2 weeks exceeded the 100KiB/s limit for the default logging.

Let's try to return to this and see if we can reduce the logspam.

This patch will be followed up with a second one to remove the old node pool

This patch tweaks the cluster and name to now be dependent on the cluster
resource rather than specifying a raw string. See [0]

[0] https://github.com/hashicorp/terraform-provider-google/issues/7851#issuecomment-730399914

Bug: T390698