prod(tf): add node pool without highthroughput logging (#2076)
This patch adds a new identical node pool to production without highthroughput logging.

We have very rarely in the last 2 weeks exceeded the 100KiB/s limit for the default logging.

Let's try to return to this and see if we can reduce the logspam.

This patch will be followed up with a second one to remove the old node pool

This patch tweaks the cluster of the node pool to be specified by the
complete cluster id e.g. projects/wikibase-cloud/locations/europe-west3-a/clusters/wbaas-2
rather than just the name.

This is actually now the format in the docs [0]

[0] https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/container_node_pool#example-usage---using-a-separately-managed-node-pool-recommended

Bug: T390698