prod(tf): remove old node pool (#2056)
* prod(tf): add node pool without highthroughput logging

This patch adds a new identical node pool to production
without highthroughput logging.

We have very rarely in the last 2 weeks exceeded the 100KiB/s
limit for the default logging.

Let's try to return to this and see if we can reduce the logspam.

This patch will be followed up with a second one to remove the old
node pool

Bug: T390698

* prod(tf): remove old node pool

Follow up to remove this old node pool that still has
high throughput logging enabled

Bug: T390698