chore(qs-updater): increase jvm heap size and k8s limits (#2266)
Analagous to #1246 and #1245 this commit naively attempts
to just increase the memory available to these processes in
in the hope that is sufficient to get things moving again.

We follow the same pattern of max heap = 0.5x request

If this is successful it may be desirable to determine
a real upperbound for how much memory is needed for the
largest batch we can contrive.

Bug: T406426