Run MW update.php script in a k8s job (#2335)
Introduce new Job and bash script to run it that triggers a MW update.

This job copies the image and ENV variables from an existing MW Backend Pod.

It also coordinates with the backend platform-api to mark the Wiki readOnly, report the new DbVersion and then make the Wiki writable again

Co-authored-by: dena <91744937+deer-wmde@users.noreply.github.com>
Co-authored-by: Thomas Arrow <thomas.arrow@wikimedia.de>
Co-authored-by: Ollie <43674967+outdooracorn@users.noreply.github.com>
Bug: T409943