local(tf): create k8s secret for botstopper imagePull
Defines a secret to be used to pull the bot stopper image.
Like our other tf secrets this needs to be stored the
engineer in tf/env/local/terraform.tfvars

Specifically this defines a Docker config Secret[0]

This was originally created following the instructions from
anubis[1] and then the encoded content exteacted by me.

After this is merged developers will need to add this secret
to their own tf/env/local/terraform.tfvars or they won't be able
to deploy to their local env.

[0] https://kubernetes.io/docs/concepts/configuration/secret/#docker-config-secrets
[1] https://anubis.techaro.lol/docs/admin/botstopper/#kubernetes
Bug: T402205