prod(tf): create k8s secret for botstopper imagePull
Defines a secret to be used to pull the bot stopper image.
Specifically this defines a Docker config Secret[0]

This was originally created following the instructions from anubis[1] and then the encoded content extracted by me.

Like our other tf secrets this needs to be stored by the deploying engineer in tf/env/staging/terraform.tfvars

Once this commit has been deployed other engineers do not need to set this variable in terraform.tfvars
due to the default. The reason this works still slightly eludes me but it seems to be a feature of
how having a default set at all prevents prompting[2] but also the change detection logic only changing
if there is a provided new value rather than the empty string which is treated as an unprovided value.

[0] https://kubernetes.io/docs/concepts/configuration/secret/#docker-config-secrets
[1] https://anubis.techaro.lol/docs/admin/botstopper/#kubernetes
[2] https://developer.hashicorp.com/terraform/language/values/variables#reference-variable-values

Bug: T402205